+++ CHANGELOG.rst
@@ -13,6 +13,8 @@ v1.7.0
 * Fixed search ignoring `exclude_unreviewed` (Thanks acidghost)
 * Add support for using pathlib paths when loading media files (Thanks reagle)
 * Remove blocklist with long dead instances
+* Add `types` parameter to notifications.
+* Document additional notification types
 * TECHNICALLY BREAKING CHANGE, but I would be quite surprised if this actually breaks anyone: Date parsing will now, when the date string is empty, return Jan. 1st, 1970 instead. This is to work around what I assume is a bug in Pleroma.
 
 v1.6.3